Fine tune -position argument
asclock option "-position -0-0" or -0+0 works better if the
default geometry (char *def_geom) is set, because now it get's
displayed at the corner and not outside of it.
A little inaccuracy is still there because the here hardcoded
borderwidth =1 can differ from the actual window manager.

Debian patch 11 (position.diff)